Move kvstore validation tests to common module
Move kvstore validation unit tests from mint/test/kvstore.rs to the
shared database/kvstore.rs module where the validation functions are
defined. This consolidates the tests with their implementation.
